add pause and resume commands
