+++ PyDodo/pydodo/__init__.py
@@ -4,4 +4,4 @@ from .request_change import (change_altitude, change_heading, change_speed,
                              change_vertical_speed)
 from .create_aircraft import create_aircraft
 from .load_scenario import load_scenario
-from .reset_simulation import reset_simulation
+from .manage_simulation import reset_simulation, pause_simulation, resume_simulation
