+++ config.yml
@@ -10,6 +10,8 @@ default:
   # Bluebird API endpoints (to be removed in favour of OpenAPI)
   status_code_aircraft_id_not_found: 404
   endpoint_reset_simulation: "reset"
+  endpoint_pause_simulation: "hold"
+  endpoint_resume_simulation: "op"
   endpoint_load_scenario: "ic"
   endpoint_create_aircraft: "cre"
   endpoint_aircraft_position: "pos"
@@ -24,8 +26,8 @@ default:
   longitude: "longitude"
   altitude: "altitude"
   ground_speed: "ground_speed"
-  vertical_speed: "vertical_speed"  
-  
+  vertical_speed: "vertical_speed"
+
   # Physical
   feet_altitude_upper_limit: 6000
   flight_level_lower_limit: 60
